+++ CHANGES.md
@@ -1,10 +1,11 @@
-## HEAD
+## [2.3.0]
+> Mar 14, 2016
 
 - Add callback support to all `fetch` methods
 - **Breakage:** Use `requestInput` property instead of `requestURL` in `requestInfo`
   middleware
 
-[HEAD]: https://github.com/mjackson/http-client/compare/latest...HEAD
+[2.3.0]: https://github.com/mjackson/http-client/compare/v2.2.1...v2.3.0
 
 ## [2.2.1]
 > Mar 13, 2016